fix: backfill script, crdb usage type (#1089)
<!-- Please make sure there is an issue that this PR is correlated to. -->

## Changes

<!-- If there are frontend changes, please include screenshots. -->

+++ lib/bolt/config/src/ns.rs
@@ -417,8 +417,8 @@ pub enum CockroachDBProvider {
 		spend_limit: u32,
 		/// CRDB Request Units.
 		/// https://www.cockroachlabs.com/docs/cockroachcloud/architecture#request-unit-ru
-		request_unit_limit: u32,
-		/// Storage limit in MiB.
+		request_unit_limit: u64,
+		/// Storage limit in GiB.
 		storage_limit: u32,
 	},
 }
